Rebase Blood Drinker System, Arachne, Oneirophage (#438)
# Description
This is a simple rebase of the Blood Drinker System, and its related
features that have been commented out and/or omitted due to its lack of
rebase. I am NOT substantially updating any of this code at this time,
outside of the barest minimum updates needed to make it run in the first
place. The reason I am doing this is that I require the Blood Drinker
system functional as a prerequisite for future features, and I will
update or refactor it when needed.

Arachne are still pending a Full Rework, but that is beyond the scope of
this PR.

# TODO

- [x] Make the code functional
- [x] Port Arachne
- [x] Uncomment Oneirophages
- [x] Re-add Oneirophage midround event

# Changelog

:cl:
- add: Arachne have been reimplemented!
- add: Oneirophages are back!

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
# Conflicts:
#	Resources/Prototypes/Chemistry/metabolizer_types.yml
#	Resources/Prototypes/Damage/containers.yml
#	Resources/Prototypes/Damage/modifier_sets.yml
#	Resources/Prototypes/Entities/Clothing/OuterClothing/armor.yml
#	Resources/Prototypes/Entities/Clothing/OuterClothing/base_clothingouter.yml
#	Resources/Prototypes/Entities/Clothing/OuterClothing/suits.yml
#	Resources/Prototypes/GameRules/events.yml
#	Resources/Prototypes/Nyanotrasen/Entities/Body/Parts/spider.yml
#	Resources/Prototypes/Nyanotrasen/Entities/Markers/Spawners/ghost_roles.yml
#	Resources/Prototypes/Nyanotrasen/Entities/Mobs/NPCs/mutants.yml
#	Resources/Prototypes/_DV/Body/Prototypes/arachne.yml
#	Resources/Prototypes/_DV/Body/Prototypes/vampiricanimal.yml
#	Resources/Prototypes/tags.yml